Comments! Lots of 'em

+++ src/ui/main_menu.rs
@@ -1,11 +1,13 @@
-// ui/main_menu.rs
+// src/ui/main_menu.rs
 
-use super::constants::{ART, TITLE};
-use super::utils::centered_rect;
-use crate::app::App;
-use crate::app_state::AppState;
-use crate::message::MessageType;
+// Import required modules and structs from other parts of the application or external crates.
+use super::constants::{ART, TITLE}; // Constants like ART and TITLE for UI.
+use super::utils::centered_rect; // Utility function for centering rectangles in the UI.
+use crate::app::App; // Main application struct.
+use crate::app_state::AppState; // Enum for managing application state.
+use crate::message::MessageType; // Enum for different types of messages.
 use ratatui::{
+    // Library for building text-based user interfaces.
     layout::{Alignment, Constraint, Direction, Layout, Margin, Rect},
     style::{Color, Modifier, Style},
     text::{Line, Span},
@@ -13,7 +15,9 @@ use ratatui::{
     Frame,
 };
 
+// Function to draw the main menu interface.
 pub fn draw_main_menu(f: &mut Frame, app: &App) {
+    // Define layout constraints for different sections of the main menu.
     let chunks = Layout::default()
         .direction(Direction::Vertical)
         .constraints(
@@ -29,6 +33,7 @@ pub fn draw_main_menu(f: &mut Frame, app: &App) {
         )
         .split(f.size());
 
+    // Render individual parts of the main menu using the layout defined above.
     render_header(f, chunks[0]);
     render_art(f, chunks[1]);
     render_title(f, chunks[2]);
@@ -37,6 +42,7 @@ pub fn draw_main_menu(f: &mut Frame, app: &App) {
     render_status(f, app, chunks[5]);
 }
 
+// Function to render the header section of the menu.
 pub fn render_header(f: &mut Frame, area: Rect) {
     let header = Paragraph::new(format!("Sharad Ratatui v{}", env!("CARGO_PKG_VERSION")))
         .style(Style::default().fg(Color::DarkGray))
@@ -45,12 +51,13 @@ pub fn render_header(f: &mut Frame, area: Rect) {
     f.render_widget(header, area);
 }
 
+// Function to render the art section of the menu.
 pub fn render_art(f: &mut Frame, area: Rect) {
     let outer_block = Block::default().style(Style::default().fg(Color::DarkGray));
     f.render_widget(outer_block, area);
 
-    let center_x = area.x + (area.width - 80) / 2;
-    let center_y = area.y + (area.height - 18) / 2;
+    let center_x = area.x + (area.width - 80) / 2; // Calculate center x for inner rectangle.
+    let center_y = area.y + (area.height - 18) / 2; // Calculate center y for inner rectangle.
     let inner_rect = Rect::new(center_x, center_y, 80, 18);
 
     let inner_block = Block::default()
@@ -64,6 +71,7 @@ pub fn render_art(f: &mut Frame, area: Rect) {
     f.render_widget(art, inner_rect);
 }
 
+// Function to render the title section of the menu.
 pub fn render_title(f: &mut Frame, area: Rect) {
     let outer_block = Block::default().style(Style::default().fg(Color::DarkGray));
     let title_outer_area = centered_rect(100, 100, area);
@@ -87,6 +95,7 @@ pub fn render_title(f: &mut Frame, area: Rect) {
     f.render_widget(title, title_inner_area);
 }
 
+// Function to render the console section of the menu.
 pub fn render_console(f: &mut Frame, app: &App, area: Rect) {
     let outer_block = Block::default().style(Style::default().fg(Color::DarkGray));
     let console_outer_area = centered_rect(100, 100, area);
@@ -120,7 +129,9 @@ pub fn render_console(f: &mut Frame, app: &App, area: Rect) {
     f.render_widget(text, console_inner_area);
 }
 
+// Function to render the interactive menu section of the main menu.
 pub fn render_menu(f: &mut Frame, app: &App, area: Rect) {
+    // Define menu items to be displayed.
     let menu_items = [
         "Start a new game",
         "Load a game",
@@ -128,6 +139,7 @@ pub fn render_menu(f: &mut Frame, app: &App, area: Rect) {
         "Settings",
     ];
 
+    // Map menu items to text lines, applying different styles to the selected item.
     let text: Vec<Line> = menu_items
         .iter()
         .enumerate()
@@ -152,7 +164,6 @@ pub fn render_menu(f: &mut Frame, app: &App, area: Rect) {
 
     let outer_block = Block::default()
         .borders(Borders::NONE)
-        // .title("Menu")
         .style(Style::default().fg(Color::DarkGray));
 
     let menu_area = centered_rect(100, 100, area);
@@ -167,7 +178,7 @@ pub fn render_menu(f: &mut Frame, app: &App, area: Rect) {
         ])
         .split(menu_area.inner(Margin {
             vertical: 1,
-            horizontal: ((area.width as u16 - text[0].width() as u16) / 2 as u16),
+            horizontal: ((area.width - text[0].width() as u16) / 2),
         }))[1];
 
     let menu = Paragraph::new(text)
@@ -176,7 +187,9 @@ pub fn render_menu(f: &mut Frame, app: &App, area: Rect) {
     f.render_widget(menu, inner_area);
 }
 
+// Function to render the status bar at the bottom of the menu.
 pub fn render_status(f: &mut Frame, app: &App, area: Rect) {
+    // Define the status message based on the current application state.
     let status_message = match app.state {
         AppState::MainMenu => "Press q to quit",
         AppState::LoadMenu => {
@@ -191,4 +204,4 @@ pub fn render_status(f: &mut Frame, app: &App, area: Rect) {
     f.render_widget(status, area);
 }
 
-// Update the draw_main_menu function to pass the app to render_status
+// Note: The proposed update to pass the 'app' to 'render_status' is already implemented in 'draw_main_menu'.
